# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "knowledge_point_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Point Ids",
            "type": "array"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resource Ids",
            "type": "array"
          },
          "task_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "user_id": {
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "user_id",
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "mode": {
            "default": "offline",
            "enum": [
              "offline",
              "external"
            ],
            "title": "Mode",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "confidence",
          "request_id"
        ],
        "title": "AskResponse",
        "type": "object"
      },
...
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "chapter": {
            "minLength": 1,
            "title": "Chapter",
            "type": "string"
          },
          "chunk_id": {
            "minLength": 1,
            "title": "Chunk Id",
            "type": "string"
          },
          "citation_id": {
            "minLength": 1,
            "title": "Citation Id",
            "type": "string"
          },
          "page_end": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "quote": {
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          },
          "resource_id": {
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          },
          "retrieval_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": -1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Score"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "citation_id",
          "chunk_id",
          "resource_id",
          "title",
          "chapter",
          "quote"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "TaskPublishRequest": {
        "additionalProperties": false,
        "properties": {
          "ai_feedback_points": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Ai Feedback Points",
            "type": "array"
          },
          "completion_criteria": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Completion Criteria",
            "type": "array"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "knowledge_point_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Knowledge Point Ids",
            "type": "array"
          },
          "prerequisite_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Ids",
            "type": "array"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Resource Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "published",
            "description": "发布状态；传 draft 可先存为草稿，稍后再发布。"
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "description",
          "task_type",
          "knowledge_point_ids",
          "resource_ids",
          "completion_criteria",
          "ai_feedback_points"
        ],
        "title": "TaskPublishRequest",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "draft",
          "published",
          "completed"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskStatusUpdateRequest": {
        "additionalProperties": false,
        "description": "任务状态流转请求，用于发布草稿或标记任务完成。",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "TaskStatusUpdateRequest",
        "type": "object"
      },
      "TaskTemplate": {
        "additionalProperties": false,
        "properties": {
          "ai_feedback_points": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Ai Feedback Points",
            "type": "array"
          },
          "completion_criteria": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Completion Criteria",
            "type": "array"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "knowledge_point_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Knowledge Point Ids",
            "type": "array"
          },
          "prerequisite_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Ids",
            "type": "array"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Resource Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "description",
          "task_type",
          "knowledge_point_ids",
          "resource_ids",
          "completion_criteria",
          "ai_feedback_points",
          "status"
        ],
        "title": "TaskTemplate",
        "type": "object"
      },
      "TaskType": {
        "enum": [
          "foundation",
          "protocol_analysis",
          "case_study",
          "innovation_challenge",
          "project_practice",
          "troubleshooting"
        ],
        "title": "TaskType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "完全独立于 Canvas 的离线优先项目骨架。",
    "title": "计算机网络 AI 助教 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/qa/ask": {
      "post": {
        "operationId": "ask_question_api_v1_qa_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "请求不符合接口契约"
          }
        },
        "summary": "Ask Question",
        "tags": [
          "qa"
        ]
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "description": "按任务状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按任务状态过滤",
              "title": "Status"
            }
          },
          {
            "description": "按六类任务类型过滤",
            "in": "query",
            "name": "task_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按六类任务类型过滤",
              "title": "Task Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "publish_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/status": {
      "patch": {
        "description": "在草稿、已发布和已完成之间流转任务状态。",
        "operationId": "update_task_status_api_v1_tasks__task_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Task Status",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}